+++ src/pcg/utils/entity_creator.cpp
@@ -60,8 +60,8 @@ void EntityCreatorImpl::handle_room_content_generation(Scene& scene, const Bound
 
     auto room = RCP.get_next_room();
 
-    for(int i = 0; i < room.size(); ++i) {
-        for(int j = 0; j < room[0].size(); ++j) {
+    for(size_t i = 0; i < room.size(); ++i) {
+        for(size_t j = 0; j < room[0].size(); ++j) {
             Entity entity = scene.create_entity();
             entity.add_component<TransformationComponent>(room_box.get_top_left().shifted_by(i, j));
             entity.add_component<GraphicsComponent>(get_color_for_room_tile(room[i][j]), Dimensions2i(1, 1));
